Added base class for DetectorInterface classes plus updated docs
