+++ src/lclstreamer/event_data_sources/generic/data_sources.py
@@ -237,3 +237,125 @@ class SourceIdentifier(DataSourceProtocol):
                 source identifier defined at initialization
         """
         return self._source_identifier
+
+class BaseDetectorInterface(DataSourceProtocol):
+    def __init__(
+        self,
+        name: str,
+        parameters: DataSourceParameters,
+        additional_info: dict[str, Any],
+    ):
+        """
+        Initializes the DetectorInterface base class
+
+        Arguments:
+
+            name: An identifier for the data source
+
+            parameters: The data source configuration parameters
+        """
+        self._name: str = name
+        extra_parameters: dict[str, Any] | None = parameters.__pydantic_extra__
+        self._call_get_data: list[tuple[str, Any, Any]] = []
+
+        if extra_parameters is None:
+            log_error_and_exit(
+                f"Entries needed by the {name} data source are not defined"
+            )
+            return  # For the type checker
+        if "psana_name" not in extra_parameters:
+            log_error_and_exit(
+                f"Entry 'psana_name' is not defined for data source {name}"
+            )
+
+        self._detector_name: str = extra_parameters["psana_name"]
+        detector_interface: Any = self._create_detector()
+
+        self.dtype: type
+        if "dtype" not in extra_parameters:
+            self.dtype = numpy.float64
+        else:
+            self.dtype = extra_parameters["dtype"]
+
+        if "psana_fields" not in extra_parameters:
+            if ":" in self._detector_name:
+                # it is a PV
+                self._call_get_data.append((self._detector_name, detector_interface, self._get_callable_with_event))
+            else:
+                log_error_and_exit(
+                    f"Entry 'psana_fields' is not defined for data source {name}"
+                )
+        else:
+            fields: list[str] | str = extra_parameters["psana_fields"]
+            det_fields: list[str] = ([fields] if isinstance(fields, str) else fields)
+            det_fields = [f.split(".") for f in det_fields]
+
+            for psana_fields in det_fields:
+                data_caller: Any = None
+                base = detector_interface
+                psana_field: str = ".".join([self._detector_name, *psana_fields])
+
+                for field in psana_fields:
+                    # Find the full name of the function we will call
+                    if hasattr(base, field):
+                        base = getattr(base, field)
+                    else:
+                        log_error_and_exit(f"Detector {base} has no parameter {field}")
+
+                if callable(base):
+                    # Check if bound method or not plus the number of args
+                    arg_number = base.__code__.co_argcount - (1 if hasattr(base, "__self__") else 0)
+                    if arg_number > 0:
+                        data_caller = self._get_callable_with_event
+                    else:
+                        data_caller = self._get_callable_with_noevent
+                else:
+                    data_caller = self._get_noncallable
+
+                data_caller = self._setup_special_fields(psana_fields, data_caller)
+
+                self._call_get_data.append((psana_field, base, data_caller))
+
+    def _setup_special_fields(self, psana_fields, data_caller):
+        return data_caller
+
+    def _get_callable_with_event(self, name, base, event):
+        return (name, numpy.array(base(event), dtype=self.dtype))
+
+    def _get_callable_with_noevent(self, name, base, event):
+        return (name, numpy.array(base(), dtype=self.dtype))
+
+    def _get_noncallable(self, name, base, event):
+        return (name, numpy.array(base, dtype=self.dtype))
+
+    def _create_detector(self, *args, **kwargs):
+        raise NotImplementedError("Derived classes have to implement their _create_detector")
+
+    def get_data(self, event: Any) -> NDArray[Any]:
+        """
+        Retrieves Detector values from a psana event
+
+        Arguments:
+
+            event: A psana1 or 2 event
+
+         Returns:
+
+            value: The retrieved data in the format of a numpy array
+        """
+        data_dict: dict[str, Any] = {}
+        name: str
+        base: Any
+        data_caller: Any
+        data = Any
+
+        for name, base, data_caller in self._call_get_data:
+            name, data = data_caller(name, base, event)
+            if isinstance(data, dict):
+                log_error_and_exit(
+                    f"Data for the psana data source {self._name} has "
+                    "the format of a dictionary! HSD detectors are not supported yet."
+                )
+            data_dict[name] = data
+
+        return data_dict
