+++ src/lclstreamer/event_data_sources/psana2/data_sources.py
@@ -6,7 +6,7 @@ from numpy.typing import NDArray
 from ...models.parameters import DataSourceParameters
 from ...utils.logging import log_error_and_exit
 from ...utils.protocols import DataSourceProtocol
-
+from ..generic.data_sources import BaseDetectorInterface
 
 class Psana2Timestamp(DataSourceProtocol):
     """
@@ -46,7 +46,7 @@ class Psana2Timestamp(DataSourceProtocol):
         return numpy.array(event.timestamp, dtype=numpy.float64)
 
 
-class Psana2DetectorInterface(DataSourceProtocol):
+class Psana2DetectorInterface(BaseDetectorInterface):
     """
     See documentation of the `__init__` function
     """
@@ -65,102 +65,16 @@ class Psana2DetectorInterface(DataSourceProtocol):
             name: An identifier for the data source
 
             parameters: The data source configuration parameters
+
+            additional_info: Contains run information etc.
         """
+        self._additional_info = additional_info
+        super().__init__(name, parameters, additional_info)
 
-        extra_parameters: dict[str, Any] | None = parameters.__pydantic_extra__
-        self._call_get_data: list[tuple[str, Any, Any]] = []
-
-        if extra_parameters is None:
-            log_error_and_exit(
-                f"Entries needed by the {name} data source are not defined"
-            )
-            return  # For the type checker
-
-        if "psana_name" not in extra_parameters:
-            log_error_and_exit(
-                f"Entry 'psana_name' is not defined for data source {name}"
-            )
-        detector_name = extra_parameters["psana_name"]
-        detector_interface: Any = additional_info["run"].Detector(
-            detector_name
+    def _create_detector(self):
+        return self._additional_info["run"].Detector(
+            self._detector_name
         )
-
-        if "psana_fields" not in extra_parameters:
-            if ":" in detector_name:
-                # it is a PV
-                self._call_get_data.append((detector_name, detector_interface, self._get_callable_with_event))
-            else:
-                log_error_and_exit(
-                    f"Entry 'psana_fields' is not defined for data source {name}"
-                )
-        else:
-            fields: list[str] | str = extra_parameters["psana_fields"]
-            det_fields: list[str] = ([fields] if isinstance(fields, str) else fields)
-            det_fields = [f.split(".") for f in det_fields]
-
-            for psana_fields in det_fields:
-                data_caller: Any = None
-                base = detector_interface
-                psana_field: str = ".".join([detector_name, *psana_fields])
-
-                for field in psana_fields:
-                    # Find the full name of the function we will call
-                    if hasattr(base, field):
-                        base = getattr(base, field)
-                    else:
-                        log_error_and_exit(f"Detector {base} has no parameter {field}")
-
-                if callable(base):
-                    # Check if bound method or not plus the number of args
-                    arg_number = base.__code__.co_argcount - (1 if hasattr(base, "__self__") else 0)
-                    if arg_number > 0:
-                        data_caller = self._get_callable_with_event
-                    else:
-                        data_caller = self._get_callable_with_noevent
-                else:
-                    data_caller = self._get_noncallable
-                self._call_get_data.append((psana_field, base, data_caller))
-
-        self.dtype: type
-        if "dtype" not in extra_parameters:
-            self.dtype = numpy.float64
-        else:
-            self.dtype = extra_parameters["dtype"]
-
-    def _get_callable_with_event(self, name, base, event):
-        return (name, numpy.array(base(event), dtype=self.dtype))
-
-    def _get_callable_with_noevent(self, name, base, event):
-        return (name, numpy.array(base(), dtype=self.dtype))
-
-    def _get_noncallable(self, name, base, event):
-        return (name, numpy.array(base, dtype=self.dtype))
-
-    def get_data(self, event: Any) -> NDArray[Any]:
-        """
-        Retrieves Detector values from a psana2 event
-
-        Arguments:
-
-            event: A psana2 event
-
-         Returns:
-
-            value: The retrieved data in the format of a numpy array
-        """
-        data_dict: dict[str, Any] = {}
-
-        for name, base, data_caller in self._call_get_data:
-            name, data = data_caller(name, base, event)
-            if isinstance(data, dict):
-                log_error_and_exit(
-                    f"Data for the psana2 data source {self._name} has "
-                    "the format of a dictionary! HSD detectors are not supported yet."
-                )
-            data_dict[name] = data
-
-        return data_dict
-
 
 class Psana2RunInfo(DataSourceProtocol):
     """
